Break cycle between notebook.ts and notebookLegacy.ts.
