+++ packages/cursorless-vscode/src/ide/vscode/notebook/notebook.ts
@@ -1,38 +1,18 @@
-import { NotebookCell, TextDocument, window } from "vscode";
-import {
-  getNotebookFromCellDocumentLegacy,
-  isVscodeLegacyNotebookVersion,
-} from "./notebookLegacy";
+import * as semver from "semver";
+import {version} from "vscode";
+import { TextDocument } from "vscode";
+import { getNotebookFromCellDocumentLegacy } from "./notebookLegacy";
+import { getNotebookFromCellDocumentCurrent } from "./notebookCurrent";
 
-/**
- * Given a document corresponding to a single cell, retrieve the notebook
- * document for the entire notebook
- * @param document The document corresponding to the given cell
- * @returns The notebook document corresponding to the notebook containing the
- * given cell
- */
+
+export function isVscodeLegacyNotebookVersion() {
+  return semver.lt(version, "1.68.0");
+}
+
 export function getNotebookFromCellDocument(document: TextDocument) {
   if (isVscodeLegacyNotebookVersion()) {
     return getNotebookFromCellDocumentLegacy(document);
+  } else {
+    return getNotebookFromCellDocumentCurrent(document);
   }
-
-  // FIXME: All these type casts are necessary because we've pinned VSCode
-  // version type defs.  Can remove them once we are using more recent type defs
-  const { notebookEditor } =
-    ((window as any).visibleNotebookEditors as any[])
-      .flatMap((notebookEditor: any) =>
-        (
-          (
-            notebookEditor.document ?? notebookEditor.notebook
-          ).getCells() as NotebookCell[]
-        ).map((cell) => ({
-          notebookEditor,
-          cell,
-        })),
-      )
-      .find(
-        ({ cell }) => cell.document.uri.toString() === document.uri.toString(),
-      ) ?? {};
-
-  return notebookEditor;
 }
