+++ packages/cursorless-vscode/src/ide/vscode/notebook/notebookCurrent.ts
@@ -1,0 +1,30 @@
+import { NotebookCell, TextDocument, window } from "vscode";
+
+/**
+ * Given a document corresponding to a single cell, retrieve the notebook
+ * document for the entire notebook
+ * @param document The document corresponding to the given cell
+ * @returns The notebook document corresponding to the notebook containing the
+ * given cell
+ */
+export function getNotebookFromCellDocumentCurrent(document: TextDocument) {
+  // FIXME: All these type casts are necessary because we've pinned VSCode
+  // version type defs.  Can remove them once we are using more recent type defs
+  const { notebookEditor } =
+    ((window as any).visibleNotebookEditors as any[])
+      .flatMap((notebookEditor: any) =>
+        (
+          (
+            notebookEditor.document ?? notebookEditor.notebook
+          ).getCells() as NotebookCell[]
+        ).map((cell) => ({
+          notebookEditor,
+          cell,
+        })),
+      )
+      .find(
+        ({ cell }) => cell.document.uri.toString() === document.uri.toString(),
+      ) ?? {};
+
+  return notebookEditor;
+}
